requirements.txt
opencv_python
torch
torchsummary>=1.5.1
tqdm
torchvision
numpy
Pillow
termcolor
ImageCaptcha